requirements.txt
aiofiles==24.1.0
ajsonrpc==1.2.0
anyio==4.8.0
argcomplete==3.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.1.dev0
autocommand==2.2.2
bottle==0.13.2
Brlapi==0.8.6
build==1.2.2
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
coverage==7.6.10
cryptography==44.0.0
cycler==0.12.1
distro==1.9.0
fastjsonschema==2.21.1
fido2==1.1.3
filelock==3.17.0
fonttools==4.56.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httptools==0.6.4
idna==3.10
ifaddr==0.2.0
inflect==7.5.0
iniconfig==2.0.0
installer==0.7.0
jaraco.classes==3.4.0
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
jeepney==0.8.0
Jinja2==3.1.5
keyring==25.5.0
kiwisolver==1.4.5
lensfun==0.3.4
libfdt==1.7.2
lit==19.1.7.dev0
louis==3.32.0
Markdown==3.7
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.10.0
meson==1.7.0
more-itertools==10.5.0
netifaces==0.11.0
nose==1.3.7
numpy==2.2.2
ordered-set==4.1.0
packaging==24.2
pillow==11.1.0
pipx==1.7.1
platformdirs==4.3.6
pluggy==1.5.0
ply==3.11
pybind11==2.13.6
pycparser==2.22
pyelftools==0.31
PyOpenGL==3.1.7
pyOpenSSL==24.3.0
pyparsing==3.1.2
pyproject_hooks==1.2.0
pyscard==2.0.10
pyserial==3.5
PySide6==6.8.2.1
pytest==8.3.4
python-dateutil==2.9.0
pyxdg==0.28
PyYAML==6.0.2
Reflector==2023.6.28.0.36.1
requests==2.32.3
SecretStorage==3.3.3
semantic_version==2.10.0
setuptools==75.2.0
setuptools-scm==8.1.0
shiboken6==6.8.2.1
shiboken6-generator==6.8.2.1
shtab==1.7.2.dev0+g61ec99d.d20241116
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.45.3
tabulate==0.9.0
TBB==0.2
termcolor==2.5.0
tldr==3.3.0
tomli==2.0.1
tqdm==4.67.1
trove-classifiers==2025.1.15.22
typeguard==4.4.1
typing_extensions==4.12.2
urllib3==2.3.0
userpath==1.9.2
uvicorn==0.34.0
uvloop==0.21.0
validate-pyproject==0.23.post1.dev0+gf45606b.d20250111
vdf==4.0
websockets==12.0
wheel==0.45.0
wsproto==1.2.0
wxPython==4.2.2
yubikey-manager==5.5.1
zeroconf==0.136.2
